refactor(wallet): split inMemoryStores into multiple files
