+++ packages/wallet/src/persistence/inMemoryStores/index.ts
@@ -1,0 +1,4 @@
+export * from './InMemoryDocumentStore';
+export * from './InMemoryCollectionStore';
+export * from './InMemoryKeyValueStore';
+export * from './inMemoryWalletStores';
